Milestone 2/M2/requirements.txt
numpy
pandas
PyYAML
requests
SQLAlchemy
psycopg2-binary
scikit-learn
docker
kafka-python
